+++ frontend-react/package.json
@@ -19,7 +19,7 @@
     "classnames": "^2.3.1",
     "dompurify": "^3.0.0",
     "downloadjs": "^1.4.7",
-    "export-to-csv": "^0.2.1",
+    "export-to-csv-fix-source-map": "^0.2.1",
     "history": "^5.3.0",
     "lodash.chunk": "^4.2.0",
     "lodash.omit": "^4.5.0",
