+++ frontend-react/src/components/FileHandlers/FileHandlerMessaging.test.tsx
@@ -6,7 +6,6 @@ import { formattedDateFromTimestamp } from "../../utils/DateTimeUtils";
 import { Destination } from "../../resources/ActionDetailsResource";
 import { conditionallySuppressConsole } from "../../utils/TestUtils";
 import { ErrorCode, ResponseError } from "../../config/endpoints/waters";
-import { FileType } from "../../hooks/UseFileHandler";
 
 import {
     RequestLevel,
@@ -16,6 +15,7 @@ import {
     ValidationErrorMessageProps,
     ValidationErrorMessage,
 } from "./FileHandlerMessaging";
+import { FileType } from "../../utils/TemporarySettingsAPITypes";
 
 // Note: following a pattern of finding elements by text (often text passed as props)
 // then asserting className. This is not ideal, and seems kinda backwards, but I couldn't
